+++ contracts/tests/ExchangeTest.sol
@@ -10,7 +10,7 @@ contract ExchangeTest {
         markPriceOracle = _markPriceOracle;
     }
 
-    function addObservation(uint256 _priceCumulative, uint64 _index) external {
+    function addObservation(uint256 _priceCumulative, uint256 _index) external {
         markPriceOracle.addObservation(_priceCumulative, _index, 0x6c00000000000000000000000000000000000000000000000000000000000000);
     }
 }
